Inject Field Calculators as Deps to Report Route
Although it seems idiomatic to have the `/report` endpoint defined as a
core endpoint, some reports may need to perform project specific calculations on
render. To support this, customization plugins can define calculator
dependencies, which will then override the dependencies supplied by this
core module. If a report specifies that it needs a calculated field but
no calculators are injected, a `NoCustomCalculatorError` will be raised
by the application.