Use Paginated Responses to Render Report Events/Instruments
Uses the new paginated responses to render report events and instruments. The `.../render/{uuid}`
endpoint has been replaced with two new endpoints: `reports/events/{uuid}` and
`reports/instruments/{uuid}`, which return paginated responses of report data constructed based
on the report definition for the report with `{uuid}`.

Calculated event and instrument data has also been reworked slightly: Any calculated event and
instrument data should be calculated in relation to the provided pages and is returned along with
paged Event/Instrument data. This has the side effect of returning ~2x the page size for reports
with calculated fields, but is logical in that calculated Event/Instrument data is returned along
with the accompanying REDCap data and not in a later page.