Add Unique Constraint on Event/Instrument Columns
Adds a unique constraint onto the `record_id`, `repeat_instance`, `event_id`, and `instrument_id` columns
in both the event and instrument tables. This combination of columns is necessarily unique for each row.
This constraint is useful for INSERT...ON CONFLICT statements where we don't know the PK id but still
want to perform efficient Upserts.